+++ libraries/strings/src/__tests__/interpolate.unit.test.ts
@@ -1,12 +1,14 @@
 import { interpolate } from '../interpolate.ts';
-import { assertEquals, describe, it } from '../../dev_deps.ts';
+import { assertEquals, assertThrows, describe, it } from '../../dev_deps.ts';
 
 describe('interpolate()', () => {
   it('when a placeholder matches a key in the dictionary, replaces the placeholder with the corresponding value', () => {
     const input = 'Hello, {name}!';
     const dictionary = { name: 'World' };
     const expectedOutput = 'Hello, World!';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
@@ -14,7 +16,9 @@ describe('interpolate()', () => {
     const input = 'Hello, {name}!';
     const dictionary = { greeting: 'Hi' };
     const expectedOutput = 'Hello, {name}!';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
@@ -22,9 +26,11 @@ describe('interpolate()', () => {
     const input = 'Hello, {name}!';
     const dictionary = { greeting: 'Hi' };
     const expectedOutput = 'Hello, name!';
+
     const output = interpolate(input, dictionary, {
       fallbackToKey: true,
     });
+
     assertEquals(output, expectedOutput);
   });
 
@@ -32,7 +38,9 @@ describe('interpolate()', () => {
     const input = '{delimiter}Hello, {name}!{delimiter}';
     const dictionary = { name: 'World', delimiter: '|' };
     const expectedOutput = '|Hello, World!|';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
@@ -40,7 +48,9 @@ describe('interpolate()', () => {
     const input = 'Hello, {}!';
     const dictionary = { name: 'World' };
     const expectedOutput = 'Hello, {}!';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
@@ -48,7 +58,9 @@ describe('interpolate()', () => {
     const input = '';
     const dictionary = { name: 'World' };
     const expectedOutput = '';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
@@ -56,7 +68,9 @@ describe('interpolate()', () => {
     const input = 'Hello, {name}!';
     const dictionary = {};
     const expectedOutput = 'Hello, {name}!';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
@@ -64,25 +78,29 @@ describe('interpolate()', () => {
     const input = 'Hello, {name_1}!';
     const dictionary = { name_1: 'World' };
     const expectedOutput = 'Hello, World!';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
-  // TODO: Relax this requirement. Any string should be usable as a placeholder.
-  it('ignores placeholders that start with a number', () => {
+  it('allows placeholders that start with a number', () => {
     const input = 'Hello, {1name}!';
     const dictionary = { '1name': 'World' };
-    const expectedOutput = 'Hello, {1name}!';
+    const expectedOutput = 'Hello, World!';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
-  // TODO: Relax this requirement. Any string should be usable as a placeholder.
-  it('ignores placeholders that include a space', () => {
+  it('allows placeholders that include a space', () => {
     const input = 'Hello, {name with space}!';
     const dictionary = { 'name with space': 'World' };
-    const expectedOutput = 'Hello, {name with space}!';
+    const expectedOutput = 'Hello, World!';
+
     const output = interpolate(input, dictionary);
+
     assertEquals(output, expectedOutput);
   });
 
@@ -95,6 +113,83 @@ describe('interpolate()', () => {
     const expectedOutput = 'Hello, World!';
 
     const output = interpolate(input, dictionary);
+
+    assertEquals(output, expectedOutput);
+  });
+
+  for (
+    const badInput of [
+      'Hello, {name!',
+      'Hello, name}!',
+      'Hello, {name{}}!',
+    ]
+  ) {
+    it('if delimiters are mismatched or nested, throws an error', () => {
+      const dictionary = { name: 'World' };
+      const throwingFn = () => interpolate(badInput, dictionary);
+
+      assertThrows(
+        throwingFn,
+        Error,
+        'Text has',
+      );
+    });
+  }
+
+  it('accepts a Map as dictionary', () => {
+    const input = '{greeting}, {planet}!';
+    const dictionary = new Map()
+      .set('greeting', 'Hello')
+      .set('planet', 'World');
+    const expectedOutput = 'Hello, World!';
+
+    const output = interpolate(input, dictionary);
+
+    assertEquals(output, expectedOutput);
+  });
+
+  it('accepts a regular expression', () => {
+    const input = '{hullo}, {hallo}!';
+    const dictionary = new Map()
+      .set(/h.llo/, 'hello');
+    const expectedOutput = 'hello, hello!';
+
+    const output = interpolate(input, dictionary);
+
+    assertEquals(output, expectedOutput);
+  });
+
+  it('does not partially match a regular expression', () => {
+    const input = '{key}, {key2}!';
+    const dictionary = new Map()
+      .set(/key/, 'value');
+    const expectedOutput = 'value, {key2}!';
+
+    const output = interpolate(input, dictionary);
+
+    assertEquals(output, expectedOutput);
+  });
+
+  it('if key is a regex, does not adapt the value to the placeholder\'s case', () => {
+    const input = '{key} {KEY} {Key}';
+    const dictionary = new Map()
+      .set(/key/i, 'value');
+    const expectedOutput = 'value value value';
+
+    const output = interpolate(input, dictionary);
+
+    assertEquals(output, expectedOutput);
+  });
+
+  it('can use a catch-all to replace unmatched delimiters', () => {
+    const input = '{Key1} {key2} {key3}';
+    const dictionary = new Map()
+      .set('key1', 'value1')
+      .set('key2', 'value2')
+      .set(/.*/, 'unmatched');
+    const expectedOutput = 'Value1 value2 unmatched';
+
+    const output = interpolate(input, dictionary, { adaptCase: true });
 
     assertEquals(output, expectedOutput);
   });
@@ -125,7 +220,9 @@ describe('interpolate()', () => {
       const input = '{greeting}, {myPlanet}! This is a {grEeting}!';
       const dictionary = { greeting: 'Hello', myPlanet: 'World' };
       const expectedOutput = 'Hello, World! This is a {grEeting}!';
+
       const output = interpolate(input, dictionary, { adaptCase });
+
       assertEquals(output, expectedOutput);
     });
   });